added results list activity

+++ MeetPoll/app/src/main/java/odb234/meetpoll/MapMarker.java
@@ -1,19 +1,24 @@
 package odb234.meetpoll;
 
 
+import android.os.Parcel;
+import android.os.Parcelable;
+
 import com.google.android.gms.maps.model.LatLng;
+
+import java.util.Map;
 
 /**
  * Created by odb234 on 7/21/16.
  */
-public class MapMarker {
+public class MapMarker{
     private LatLng latLng;
     private String address;
     private String id;
-    private double rating;
+    private float rating;
     private String name;
 
-    public MapMarker(LatLng ll, String add, String newId, double r, String n){
+    public MapMarker(LatLng ll, String add, String newId, float r, String n){
         latLng = ll;
         address = add;
         id = newId;
@@ -33,7 +38,7 @@ public class MapMarker {
     public String getName(){
         return name;
     }
-    public double getRating(){
+    public float getRating(){
         return rating;
     }
     public void setLatLng(LatLng ll){
@@ -45,7 +50,7 @@ public class MapMarker {
     public void setId(String id){
         this.id = id;
     }
-    public void setRating(double r){
+    public void setRating(float r){
         this.rating = r;
     }
     public void setName(String n){
